worker/requirements.txt
redis==3.5.3
rq==1.5.2
kafka-python==2.0.2
vosk==0.3.14